requirements.txt
celery==4.1.0
Django==1.11.7
django-allauth==0.34.0
django-crispy-forms==1.7.0
django-imagekit==4.0.1
djangorestframework==3.7.3
Pillow==4.3.0
redis==2.10.6
requests==2.18.4
django-imagekit==4.0.1
django-notifications-hq
django-ckeditor